Criado as classes Conta e Banco
  Na criação da classe Banco foi aplicado o conceito de data class.
