+++ digionebank/src/main/kotlin/one/digitalinnovation/bank/Conta.kt
@@ -1,0 +1,15 @@
+package one.digitalinnovation.bank
+
+import java.math.BigDecimal
+
+class Conta (
+    val agencia: String,
+    val numero: String,
+    val saldo: BigDecimal,
+){
+    fun deposito(valor: BigDecimal){
+    }
+
+    fun saque(valor: BigDecimal){
+    }
+}
