+++ digionebank/src/main/kotlin/one/digitalinnovation/bank/Main.kt
@@ -1,0 +1,12 @@
+
+package one.digitalinnovation.bank
+
+
+fun main(){
+    val gabriel = Pessoa()
+    println(gabriel.nome)
+    println(gabriel.cpf)
+    println(gabriel.Endereco().rua)
+    Animal(tipo="Mamífero")
+
+}
