+++ digionebank/src/main/kotlin/one/digitalinnovation/bank/Pessoa.kt
@@ -31,12 +31,3 @@ class Pessoa() {
     }
 }
 
-
-fun main() {
-    val gabriel = Pessoa()
-    println(gabriel.nome)
-    println(gabriel.cpf)
-    println(gabriel.Endereco().rua)
-    Animal(tipo="Mamífero")
-}
-
